[p-34] Add dealloc

See merge request pi-os/pi-os!1

+++ src/pi/rts/memoryManagement.h
@@ -22,7 +22,7 @@ enum PageBits {
 };
 
 //Value *newValue(size_t size);
-int alloc(size_t pages);
+int8_t alloc(size_t pages);
 void init();
 
 extern size_t HEAP_START;
